lion/pyproject.toml
[project]
name = "lion"
version = "0.1.0"
description = ""
authors = [
    {name = "김예은",email = "lpdms12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "lion", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
